+++ changelog.md
@@ -209,3 +209,4 @@ Accidental version bump. No changes.
 
 * Remove ambiguous characters from passwords
 * Base login email field on user email field
+* Password reset functionality